+++ src/biz/bokhorst/xprivacy/XBinder.java
@@ -61,7 +61,9 @@ public class XBinder extends XHook {
 		"iphonesubinfo",
 		"iphonesubinfo_msim",
 		"sip",
-		"nfc"
+		"nfc",
+		"bluetooth",
+		"bluetooth_manager"
 	});
 	// @formatter:on
 
@@ -104,15 +106,15 @@ public class XBinder extends XHook {
 		"android.telephony.TelephonyManager",
 		"android.app.ApplicationPackageManager",
 		"android.telephony.TelephonyManager",
-		"android.telephony.TelephonyManager",
+		"android.telephony.MSimTelephonyManager",
 		"android.view.WindowManagerImpl,android.view.WindowManagerGlobal,android.view.ViewRootImpl,android.view.View,android.view.Display",
 		"android.net.wifi.WifiManager",
 		"android.net.sip.SipManager",
 		"android.telephony.SmsManager",
 		"android.nfc.NfcActivityManager,android.nfc.NfcAdapter",
 		"android.appwidget.AppWidgetManager,android.appwidget.AppWidgetHost",
-		"com.android.server.BluetoothManagerService,android.bluetooth.BluetoothSocket",
-		"android.bluetooth.BluetoothManager,android.bluetooth.BluetoothAdapter,android.bluetooth.BluetoothDevice",
+		"com.android.server.BluetoothManagerService,android.bluetooth.BluetoothAdapter,android.bluetooth.BluetoothSocket",
+		"android.bluetooth.BluetoothManager,android.bluetooth.BluetoothAdapter,android.bluetooth.BluetoothDevice,android.bluetooth.BluetoothPan",
 		"android.hardware.input.InputManager",
 		"android.hardware.SystemSensorManager",
 		"android.hardware.usb.UsbManager"
@@ -122,7 +124,7 @@ public class XBinder extends XHook {
 	// @formatter:off
 	// Forbidden classes
 	public static List<String> cBlackClassName = Arrays.asList(new String[] {
-		"java.lang.reflect.Method.invokeNative"
+		"java.lang.reflect.Method"
 	});
 	// @formatter:on
 
@@ -146,19 +148,25 @@ public class XBinder extends XHook {
 			"android.content.BroadcastReceiver$PendingResult",
 			"android.content.ContentResolver",
 			"android.hardware.SensorManager",
+			"android.media.MediaPlayer",
+			"android.widget.TextView",
+			"android.os.Looper",
 			"android.os.StrictMode$AndroidBlockGuardPolicy",
 			"com.android.internal.app.ResolverActivity",
+			"com.android.internal.os.RuntimeInit",
 			"com.android.internal.os.RuntimeInit$UncaughtHandler",
 		},
 		new String[] { // ClipboardManager
 		},
 		new String[] { // ConnectivityManager
 			"android.app.ActivityThread",
+			"android.net.VpnService",
 		},
 		new String[] { // ContentProvider
 		},
 		new String[] { // LocationManager
 			"android.location.Geocoder",
+			"android.telephony.cdma.CdmaCellLocation",
 		},
 		new String[] { // TelephonyManager
 		},
@@ -166,19 +174,26 @@ public class XBinder extends XHook {
 		},
 		new String[] { // PackageManager
 			"android.app.ActivityThread",
+			"android.app.ContextImpl$TctExtContextImpl",
 			"android.app.LoadedApk",
 			"android.app.ResourcesManager",
 			"android.content.res.Resources",
 			"android.content.thm.ThemeIconManager",
+			"android.hardware.SensorManager",
 			"android.nfc.NfcAdapter",
+			"com.android.internal.app.ResolverActivity$ResolveListAdapter",
 		},
 		new String[] { // TelephonyManager
 		},
 		new String[] { // TelephonyManager
 		},
 		new String[] { // WindowManager
+			"android.app.Activity",
+			"android.app.ActivityThread",
+			"android.app.KeyguardManager",
 			"android.app.KeyguardManager$KeyguardLock",
 			"android.hardware.LegacySensorManager",
+			"android.widget.PopupWindow",
 		},
 		new String[] { // WifiManager
 		},